refactor: Moving issue publishing to detector tool instead of in reporting.
